requirements.txt
streamlit
pillow
diffusers
transformers
accelerate
scipy
safetensors